Update terser to ^5.14.0
Summary:
This update includes the removal of a direct dependency on `source-map@0.7.x`, in particular the problematic (especially for React Native) `fetch`-based browser/node environment detection described in https://github.com/terser/terser/pull/1164.

This was the last item blocking us from deprecating and replacing `metro-minify-uglify`, which uses the long-depreceted [`uglify-es`](https://www.npmjs.com/package/uglify-es).

### Follow-up
I'll follow up with adding `metro-minify-terser` to `metro`'s dependencies and communicating the deprecation of `metro-minify-uglify` in OSS. Then we'll be able to switch the default and remove `metro-minify-uglify` from the deps in a release cycle or two.

Reviewed By: motiz88

Differential Revision: D36098090

fbshipit-source-id: 415431f17e8048bcd1eccd80c3faa6bbc2283cfb