Bootstrap metro-react-native-interop-tools
Summary:
We’ve created a new package for interop between JS and native code.
The main focus now is the compatibility checker, we have this tool internally, but the current implementation of the tool is quite strongly tied to Meta’s infrastructure. We’d like to explore reworking this tool to eventually make it available to the larger ecosystem.
But the package can evolve to include more than just a compatibility checker. Maybe we can merge some older React Native codegen packages into it and have a complete interop solution based on Metro.

Files added:
	.npmignore
	BUCK
	package.json
	README.md
	src/index.js - the entrypoint of the package

Reviewed By: GijsWeterings

Differential Revision: D37815755

fbshipit-source-id: ddd0142e671467b9a63e395e29fd438629216e72