metro-react-native-interop-tools added nullable comparison
Summary:
Added the nullable type annotations. If both types are nullable I will compare the types they cover, if just one type is nullable, for the left type, I will remove the wrapper and compare those. But if the right is nullable and the left is not, that leads to an incompatibility:
```
export interface Spec extends TurboModule {
   +setKeepScreenOn: (screenOn: boolean) => void;
}
```
                                  |
                                  V
```
export interface Spec extends TurboModule {
   +setKeepScreenOn: (screenOn: ?boolean) => void;
}
```

Reviewed By: GijsWeterings

Differential Revision: D38981392

fbshipit-source-id: 315aba5798107bbe139a1f62a9c16a097d3d136b