metro-react-native-interop-tools added object comparison
Summary: It is safe to add or remove optional properties in the object. In this case, we expect that undefined values in case we remove the property are already handled by the native code. Read-only objects will be treated as normal objects, they will be unwrapped from the generic type and be compared as normal objects. Same with inexact objects, in this case, native will not read/provide the additional properties from inexact objects if they are given.

Reviewed By: GijsWeterings

Differential Revision: D39391004

fbshipit-source-id: 2d66b38875218fad2dfedbf6e29b66bb59bc5c79