Drop support for Watchman <=4.9.0, remove dead code
Summary:
We have various capability checks across the Watchman watcher and crawler at the moment, checking for features that have been available in Watchman for a long time (before 2017 in most cases).

This drops support for any version of Watchman too old to support the features we prefer to use, by asserting on specific capabilities and removing the fallback code paths.

If the capability check fails we'll still seamlessly fall back to a zero-dependency option via the "node" crawler+watcher.

Changelog:
[Breaking]: Drop support for old (pre-CalVer) Watchman versions

Reviewed By: jacdebug

Differential Revision: D39771009

fbshipit-source-id: 2d0f84157d8daf0d7337d535be28cd92bdf048e1