Support loading source URLs in inspector-proxy
Summary:
The path to a source file may be a bundler service URL, rather than a filesystem path. This change makes the inspector-proxy detect when the source path is a URL, and fetch the source over HTTP (so long as the URL is on localhost).

This change also includes a bit of reorganizing: the message-intercepting code is now responsible for sending the response, as some interceptions generate a synchronous response, some generate an asynchronous response, and some generate no response at all.

As with errors when loading a file, a failure to load is sent to the client.

Reviewed By: robhogan

Differential Revision: D39629274

fbshipit-source-id: a205a015bff68190448f28350d481304ad207556