Ownership added oncall for react_native based on file prediction method
Summary:
== Context
[Meta 2022 JavaScript BE Goals - Ownership Coverage](https://fb.workplace.com/groups/untypedjs/permalink/609524900352718/)

== This Diff

The Ownership team did some analysis and found out that only your team owns some of the files in the directory and we believe that your oncall maybe the best owner for the other files in the same directory.

If you're the correct owner, please **accept the diff**. Also feel free to **commandeer this diff** to your own need.
If you're not the correct owner, please **leave a comment** stating the reason why you're not the correct owner, and reject the diff.

For other concerns, please create a post in this group: [Ownership Prediction Feedback](https://fb.workplace.com/groups/1525535750953180).

Reviewed By: jacdebug

Differential Revision: D40702064

fbshipit-source-id: 797d71275f652ac2163c9453d2242e1defe5c535