Clean up unused cache functions from `metro-file-map`
Summary:
Cleanup some leftovers from `metro-file-map`'s main class cache management, since we moved the bulk of this out with the `cacheManagerFactory` abstraction.

The static `getCacheFilePath()` is particularly unhelpful because `DiskCacheManager` doesn't use it, so it isn't even correct - I should've removed it when we added the separate cache manager.

These functions are unused within Metro and `metro-file-map` is still marked experimental, so removing them is semver non-breaking.

Reviewed By: jacdebug

Differential Revision: D40635935

fbshipit-source-id: c5f807ad6ef987f6a760223e7e21fa08381671eb