Remove Meta-specific ModuleGraph code from open source
Summary:
Changelog: Internal

The `metro/src/ModuleGraph` directory contains a large amount of code that is used only in Meta's internal integration of Metro and Buck. We're looking to delete/simplify some of this code and use more of Metro's public API instead. Here, we take the first step which is to remove it from the Metro open source repo.

Reviewed By: huntie

Differential Revision: D41628261

fbshipit-source-id: 8b806b0a7171a2d48675ceea8e0713b5ae3ecabc